requirements.txt
streamlit==1.25.0
tensorflow==2.12.0
numpy==1.24.3
pillow